fix: git sign email
For github to recognize the commit as a verified commit the email of the
signature must match the committer email. Otherwise it shows as
"unverified" even though the signing pubkey is present in github
settings.

Also it makes more sense to not use my sub-mail for commits because the
commit is bound to my identity and not my github email.