cfg: let units with missing secret config fail explicitly
Sometimes sops-nix doesnt correctly decrypt the secret templates used by these units.
This way they explicitly tell that they failed because of that.

The reason for sops-nix probably is gnupg locking most of the time.
Need some workaround for that. Maybe find and cat lock files to get the uid, check with ps if sth is active and rm lock otherwise