fix: arrstack: quality profiles
recyclarr keeps adding rules/scores to custom formats with each run instead of creating the defined state which after some runs fucks up the scoring.
This seems to be intended behavior.
The reset_unmatched_scores setting is supposed to mitigate this but it seems not to e enough.
My research showed that there might be some more issues:
- includes and merging settings if defined multiple times
- the generated json settings from the nixos module might be a problem

-> should remove the templates and create the configs in code instead
-> maybe not use recyclarr nixos module for the config file